pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.14.0",
    "asyncpg==0.30.0",
    "fastapi==0.115.6",
    "passlib[bcrypt]==1.7.4",
    "psycopg==3.3.2",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.12",
    "sqlalchemy==2.0.36",
    "uvicorn[standard]==0.30.6",
]
